move /ping and /_stat to HTTPService
