+++ examples/userservice.py
@@ -3,12 +3,8 @@ import json
 
 from vyked import Host, HTTPService, TCPService, get, post, api, publish, Request, Response
 
-REGISTRY_HOST = '127.0.0.1'
 REGISTRY_PORT = 4500
-
-REDIS_HOST = '127.0.0.1'
 REDIS_PORT = 6379
-
 USER_HOST = '127.0.0.1'
 USER_HTTP_PORT = 4501
 USER_TCP_PORT = 4502
