+++ vyked/__init__.py
@@ -5,8 +5,8 @@ __all__ = ['Host', 'TCPServiceClient', 'TCPService', 'HTTPService', 'HTTPService
 
 from .host import Host  # noqa
 from .services import (TCPService, HTTPService, HTTPServiceClient, TCPServiceClient)  # noqa
-from .decorators.http import (get, post, head, put, patch, delete, options, trace)  # noqa
-from .decorators.tcp import (api, request, subscribe, publish, xsubscribe, apideprecated)  # noqa
+from .services import (get, post, head, put, patch, delete, options, trace)  # noqa
+from .services import (api, request, subscribe, publish, xsubscribe, apideprecated)  # noqa
 from .registry import Registry  # noqa
 from .utils import log  # noqa
 from .exceptions import RequestException, VykedServiceError, VykedServiceException  # noqa
