+++ vyked/registry_client.py
@@ -22,7 +22,6 @@ def _retry_for_exception(_):
 
 
 class RegistryClient:
-
     def __init__(self, loop, host, port, ssl_context=None):
         self._loop = loop
         self._port = port
@@ -78,9 +77,10 @@ class RegistryClient:
            strategy=[0, 2, 4, 8, 16, 32])
     def connect(self):
         self._transport, self._protocol = yield from self._loop.create_connection(partial(get_vyked_protocol, self),
-                                                                                  self._host, self._port, ssl=self._ssl_context)
+                                                                                  self._host, self._port,
+                                                                                  ssl=self._ssl_context)
         yield from self.conn_handler.handle_connected()
-        self._pinger = TCPPinger('registry', self._protocol, self)
+        self._pinger = TCPPinger(self._host, self._port, 'registry', self._protocol, self)
         self._pinger.ping()
         return self._transport, self._protocol
